LegalMate_AI-BD/requirements.txt
fastapi
langchain_chroma
langchain_cohere
langchain_community
langchain_core
langchain_google_genai
langchain_tavily
langchain_text_splitters
pydantic
pydantic_settings
python-dotenv
uvicorn
gunicorn 